Refactor code into view mixins
close #275
